+++ src/lib/__tests__/commands/scan.spec.ts
@@ -154,6 +154,6 @@ describe("scan command", () => {
 
         action({verifyUiSync: true})
 
-        expect(mockScanStorage).toHaveBeenCalledWith("/fake/root", {verifyUiSync: true})
+        expect(mockScanStorage).toHaveBeenCalledWith("/fake/root", {verifyUiSync: true, showWarnings: true})
     })
 })